Action graph building: add some missing edges
e.g. if `a` depends on `b`, and `a` is getting removed but `b`
installed, we did not constrain the removal to happen first.

This could be meaningful in the case where it's an optional dependency
and `b` wasn't installed beforehand: without this constraint it
would be possible to get through a state where `a` was installed
together with `b`, but without consideration for the depopt.