More filtering of "available compilers"
The way the repo is currently done, dev ocaml packages are artificially made
unavailable indirectly, my making them depend on the unavailable `ocaml-beta`
package.

This patch doesn't do a complete installability check, which would be way too
expensive, but (over-)approximates the fact that direct dependencies of the
compiler are available.